Moved `convertSecondsToTimestamp` into `video.ts`
