+++ client/src/main/scripts/components/TrackSlider.tsx
@@ -1,6 +1,6 @@
 /*
  * MyTake.org website and tooling.
- * Copyright (C) 2018 MyTake.org, Inc.
+ * Copyright (C) 2018-2020 MyTake.org, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -20,7 +20,7 @@
 import * as React from "react";
 import RangeContainer from "./RangeContainer";
 import { RangeType, StateAuthority, TimeRange } from "./Video";
-import { convertSecondsToTimestamp } from "../utils/functions";
+import { convertSecondsToTimestamp } from "../common/video";
 
 export interface TrackSliderEventHandlers {
   onAfterRangeChange: (
