+++ client/src/main/scripts/utils/functions.ts
@@ -17,45 +17,6 @@
  *
  * You can contact us at team@mytake.org
  */
-export function convertSecondsToTimestamp(totalSeconds: number): string {
-  let truncated = totalSeconds | 0;
-
-  const hours = Math.floor(truncated / 3600);
-  const HH = Math.floor(truncated / 3600).toString();
-
-  truncated %= 3600;
-
-  const seconds = truncated % 60;
-  const SS = zeroPad(seconds);
-
-  const minutes = Math.floor(truncated / 60);
-
-  let MM;
-  if (hours > 0) {
-    MM = zeroPad(minutes);
-    return HH + ":" + MM + ":" + SS;
-  } else if (minutes > 0) {
-    MM = minutes.toString();
-    return MM + ":" + SS;
-  } else {
-    return "0:" + seconds.toString();
-  }
-}
-
-function zeroPad(someNumber: number): string {
-  if (someNumber == 0) {
-    return "00";
-  }
-
-  let twoDigitStr: string;
-  if (someNumber < 10) {
-    twoDigitStr = "0" + someNumber.toString();
-  } else {
-    twoDigitStr = someNumber.toString();
-  }
-  return twoDigitStr;
-}
-
 export function slugify(text: string): string {
   return text
     .toLowerCase()
